+++ app/src/main/java/com/mm/nexttasks/db/entities/Task.kt
@@ -6,11 +6,11 @@ import java.util.Date
 
 @Entity
 data class Task (
-    @PrimaryKey(autoGenerate = true) val taskId: Int,
+    @PrimaryKey(autoGenerate = true) val taskId: Long,
     val title: String,
-    val taskListId: Int,
-    val priorityId: Int?,
-    val categoryId: Int?,
+    val taskListId: Long,
+    val priorityId: Long?,
+    val categoryId: Long?,
     val isDone: Boolean,
     val cardColor: Int?,
     val term: Date?,
